requirements.txt
alembic==1.6.0
click==8.0.1
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
flask_marshmallow==0.14.0
marshmallow-sqlalchemy==0.26.1
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.26.0
SQLAlchemy==1.4.23
Werkzeug==1.0.1
itsdangerous==1.1.0
flask-restx==0.5.1
psycopg2-binary==2.8.5
pika==1.2.0